+++ ReleaseNotes.md
@@ -1,3 +1,14 @@
+## 0.12.0
+- [core-Breaking Change] hash literal now using "\#" instead of "##"
+- [all] upgrade project to netcoreapp2.0
+- [core] if statement now support equals to boolean literal(== true/false)
+- [core] sdmap can now live with error close curly brace } harmony.
+- [core] sdmap can now support close curly brace } litera, using "\}"
+- [ext] add SdmapContext
+- [ext] rename ISqlEmiter to ISdmapEmiter
+- [ext] rename EmitSql in ISqlEmiter to Emit
+- [ext] not rely on System.Threading.Thread package.
+
 ## 0.11.5
 - [core] support double hash(##) emit single hash(#).
 - [vstool] upgrade to use 0.11.5 core version.